requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
aioping==0.4.0
mac-vendor-lookup==0.1.12

schedule==1.2.0
requests==2.31.0
python-dotenv==1.0.0
opencv-python==4.8.1.78
numpy==1.24.3
wmi
psutil
pynput
Pillow
cryptography
pywin32